style: update project styling and add new fonts and styles to homepage
This commit includes changes to the project styling, including the addition of new fonts and styles to the homepage. The `globals.scss` file was renamed to `index.scss` and updated to include imports for the new font files, as well as the `resets.scss`, `base.scss`, `functions.scss`, `layouts.scss`, `mixins.scss`, and `variables.scss` files. The `page.module.scss` file was also updated to import styles from the new `homepage.module.scss` file.