style(layout.tsx): remove unused font declaration and commented code style(page.tsx): add lorem ipsum text to h1 tag style(_base.scss): add font and color variables to body tag style(_fonts.scss): import google fonts style(_mixins.scss): add font mixin style(_variables.scss): add font and color variables, media queries, and text sizes
The changes were made to improve the styling of the application. The changes include adding font and color variables, media queries, and text sizes to the application. Unused font declarations and commented code were removed. A lorem ipsum text was added to the h1 tag. A font mixin was added to the _mixins.scss file. Google fonts were imported in the _fonts.scss file.