feat(crew): add CrewPage component and its styles
Add a new component called CrewPage to the app. This component is a simple
main element with the text "CrewPage" and a class name of "main". Also, add
a new SCSS file to the app/styles/pages directory called crewpage.module.scss.
This file contains a class called "main" that uses the "container" mixin to
center the content horizontally and vertically.