feat(crew): add CrewCard component to display crew data on CrewPage
This commit adds a new component, CrewCard, to display crew data on the CrewPage. The component receives a Crew object as a prop and displays the name of the crew member. The CrewPage fetches data from the server and maps over the array of Crew objects to render a CrewCard for each one. The styles for the CrewCard are also added.